requirements.txt
taipy-gui>=4.0.0
fastf1>=3.1.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
